Increase test coverage to 84% and fix static analysis issues
- Add comprehensive unit tests for all modules, widgets, DB class,
  helper functions, and downloads page (258 tests, 401 assertions)
- Add mock DB infrastructure (MockDbHandler, MockResult) for testing
  procedural module files without a real database connection
- Fix all 13 PHPStan level 8 errors (type casts, null safety, argument types)
- Add Psalm baseline for 58 existing legacy code issues
- Suppress mail() warnings in test environment with @ operator
- Guard sub() function declaration to prevent redeclaration on multiple includes
- Update .gitignore for coverage-report directory

+++ pdl-inc/pdl_functions.inc.php
@@ -282,7 +282,7 @@ function seiten_render_limited(int $currentPage, int $totalPages, int $maxVisibl
     $range = seiten_calc_range($currentPage, $totalPages, $maxVisible);
     $output = '';
 
-    $start = max(1, $currentPage - $range['before']);
+    $start = (int) max(1, $currentPage - $range['before']);
     for ($j = $start; $j < $currentPage; $j++) {
         $output .= seiten_page_link($file, $j, $link, false);
     }
